// frontend/src/api/generated/models/userResponseRole.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * 途排智策 API
 * 途排智策前后端分离 MVP API
 * OpenAPI spec version: 0.1.0
 */

export type UserResponseRole = typeof UserResponseRole[keyof typeof UserResponseRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserResponseRole = {
  admin: 'admin',
  scheduler: 'scheduler',
  approver: 'approver',
  viewer: 'viewer',
} as const;